完成2.2.4版本开发

+++ app/src/main/java/com/li/videoapplication/data/model/response/ConfirmOrderEntity.java
@@ -10,9 +10,9 @@ public class ConfirmOrderEntity extends BaseResponseEntity {
 
     private String order_id;
 
-    private int residue_coin;
+    private float residue_coin;
 
-    public int getResidue_coin() {
+    public float getResidue_coin() {
         return residue_coin;
     }
